+++ docs/changelog.md
@@ -16,6 +16,7 @@ Click on the version in the following list to see all changes since the previous
 * test: improve test coverage and suppress unused warnings
 * test: include performance test suite for Jira Data Center
 * test: add markers to session list and detail view to identify pages
+* feat: copy session link with a button on the share session screen (#62)
 
 ### [3.12.0] - 2019-04-05